mcp-client/pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.49.0",
    "mcp>=1.6.0",
    "pdfminer-six>=20250416",
    "python-dotenv>=1.1.0",
]
